refactoring softnet
Refactors softnet sampler to move the `Statistic` enum out into its
own module.
